refactor: ListItem, ListSection 컴포넌트 수정
