+++ src/components/list-section/list-item/ListItem.tsx
@@ -1,7 +1,16 @@
 import styles from "./styles.module.scss";
 import Link from "next/link";
 import Image from "next/image";
-import { ListItemProps } from "../ListSection";
+
+export interface ListItemProps {
+  id?: string;
+  name?: string;
+  title?: string;
+  description?: string;
+  uri?: string;
+  images: { url: string; width: number; height: number }[];
+  type?: string;
+}
 
 export default function ListItem({
   id,
@@ -16,17 +25,19 @@ export default function ListItem({
       <Link className={styles["list-item-link"]} href={`/${type}/${id}`}>
         <div className={styles["list-item-album"]}>
           <Image
-            src={images?.[0]?.url}
-            width={images?.[0]?.width || 50}
-            height={images?.[0]?.height || 50}
+            src={
+              images?.[0]?.url || "https://dummyimage.com/145x145/ccc/fff.png"
+            }
+            width={images?.[0]?.width || 145}
+            height={images?.[0]?.height || 145}
             alt="album"
             className={styles["list-item-album-image"]}
           />
         </div>
-        {name && <div className={styles["list-item-title"]}>{name}</div>}
-        {description && (
-          <div className={styles["list-item-description"]}>{description}</div>
-        )}
+        <div className={styles["list-item-title"]}>{name || "undefined"}</div>
+        <div className={styles["list-item-description"]}>
+          {description || "undefined"}
+        </div>
       </Link>
     </li>
   );
